Major work on opcodes, up to 8xyE
Types are still a bit of a mess. Stack is u16 and sp is u16.
Might be possible this is invalid in certain situations.
Significant progress on implementing opcodes with match statements.
Better solution might be to wrap it up in separate functions.
Definitely want to move it out of cpu.rs.