Decode & execute opcodes 2 times for each loop. Decrement timers at 60Hz
We decrement the delay_timer & sound_timer at 60Hz instead of the previous
hacky attempt. This fixes a lot of compatibility issues with games.
Redraw is still a tad fast in some games, and a bit slow for others.
Some games operate better with a higher opcode decode count per loop..